requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.7.2
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.1.0
authy==2.2.6
autopep8==2.0.4
Babel==2.13.0
backcall==0.2.0
backoff==2.2.1
billiard==4.1.0
boto==2.49.0
boto3==1.35.8
botocore==1.35.8
cachetools==5.3.1
celery==5.3.4
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.6.1
cron-descriptor==1.4.5
cronitor==4.7.1
cryptography==41.0.6
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.0.7
dj-database-url==2.1.0
dj-inmemorystorage==2.1.0
dj-rest-auth==5.0.1
dj-stripe==2.8.2
Django==4.2.6
django-allauth==0.57.0
django-appconf==1.0.5
django-auditlog==2.3.0
django-braces==1.15.0
django-cacheops==7.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-currentuser==0.6.1
django-filter==23.3
django-fsm==2.8.1
django-fsm-admin==1.2.5
django-js-asset==2.1.0
django-jsonfield==1.4.1
django-loginas==0.3.11
django-model-utils==4.3.1
django-money==3.3.0
django-mptt==0.15.0
django-mptt-urls==2.0.4
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.2.0
django-redis==5.4.0
django-rest-framework-social-oauth2==1.1.0
django-separatedvaluesfield==0.6.0
django-ses==3.5.0
django-smtp-ssl==1.0
django-sslserver==0.22
django-storages==1.14.2
django-taggit==4.0.0
django-timezone-field==6.0.1
django-user-agents==0.4.0
djangorestframework==3.14.0
docutils==0.20.1
drf-extra-fields==3.7.0
drf-nested-routers==0.93.4
drf-yasg==1.21.7
easypost==8.1.1
exceptiongroup==1.2.0
exponent-server-sdk==2.0.0
factory-boy==3.3.0
Faker==19.6.2
filetype==1.2.0
flake8==6.1.0
flower==2.0.1
fonttools==4.43.1
freezegun==1.2.2
funcy==1.18
future==0.18.3
google-api-core==2.19.2
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-i18n-address==3.1.0
googleapis-common-protos==1.65.0
greenlet==3.0.3
h11==0.14.0
httplib2==0.22.0
hubspot-api-client==8.1.1
humanize==3.13.1
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonfield==3.1.0
jwcrypto==1.5.6
kiwisolver==1.4.5
kombu==5.3.2
lazy_loader==0.4
MarkupSafe==2.1.3
matplotlib==3.8.0
mccabe==0.7.0
mock==5.1.0
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
nose==1.3.7
nose-pattern-exclude==0.1.3
nulltype==2.3.1
numpy==1.26.0
oauth2client==4.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pbr==5.11.1
pep8==1.7.1
pexpect==4.8.0
phonenumbers==8.13.22
pickleshare==0.7.5
Pillow==10.0.1
plaid-python==16.0.0
platformdirs==3.11.0
pluggy==1.3.0
postmarker==1.0
prometheus_client==0.20.0
prompt-toolkit==3.0.39
proto-plus==1.24.0
protobuf==5.27.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
py==1.11.0
py-moneyed==3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pypng==0.20220715.0
PySocks==1.7.1
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
python-crontab==3.2.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
rudder-sdk-python==2.0.2
s3transfer==0.10.2
scikit-image==0.22.0
scipy==1.11.3
selenium==4.17.2
sentry-sdk==1.31.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sorl-thumbnail==12.10.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
stripe==4.3.0b3
sympy==1.12
text-unidecode==1.3
tifffile==2023.9.26
tomli==2.0.1
tornado==6.3.3
trio==0.24.0
trio-websocket==0.11.1
typed-ast==1.5.5
typing_extensions==4.9.0
tzdata==2024.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.2
user-agents==2.2.0
uWSGI==2.0.22
vine==5.0.0
wcwidth==0.2.8
webdriver-manager==4.0.1
wsproto==1.2.0
zipp==3.17.0